Important: Unique Cardinal moved from Foot to Algebra
When I first imagined a need for the Cardinal it was just a soft type intended to allow some who wanted to avoid bringing in an imaginary scalar part to do so. One guy on our west coast research team was toying with the idea of justifying every expansion. Since then, I've finally figured out what I want to do with the Cardinal. I intend to use it to describe the units of the scalar field over which an algebra is defined. That means uniqueness is appropriate for algebras and the Foot can simply track them for convenience.