refactored the interfaces in cladosF
The Divisible interfaces contained more conceptual material than should be done if one wants to stick to the spirit of the purpose of interfaces. So... I split them into three pieces and that allowed the two Divisible interfaces to collapse into one.

The portions of the old Divisible interfaces that were sensitive to float and doubles in parameters and results have shifted to Normalizable and Scalable interfaces.

This alteration should not change the outwardly visible behavior of the DivFields.