lots of small-ish changes that don't alter outward behaviors
besides the obvious formatting change, there are also XML output changes and some small math changes. For example, the signflip short is converted at the last moment when filling product cells from (0,1) to (1, -1) so the Math.pow() function can be avoided. The result is the same when working out product cells, though.