works for me
If you are going to build your own, alter the classpath in each target to include where you put JUnit. Someday I will adjust them here to something less direct, but it isn't my top priority right now. I'll have to update to a more recent JUnit too, so that's the real cause of the delay for me.