Mostly modernizing nyads
Took some time to check to see if I got all the short integers converted
to regular integers for blade indexes, though. The test for that was to
try for 15 generator algebras. It works, but with 1G entries in the
product table it is close to blowing the default heap size limit. 16
generators definitely does it... and likely takes more than 1 hour to
generate an algebra on my machine.