Shifting the coefficients toward MonaAbstract
Tests pass at this point, but the shift is incomplete. All coeffs have
been moved into the new Scale<Blade, DivField> object and monad methods
rewritten to point there. In the near future, though, many of those
methods will shift up to the super class which will become much less
abstract with each step.

The goal is to get them all shifted to the super class and do away with
children of MonadAbstract. If that happens, MonadAbstract will be
renamed 'Monad'.