monad methods lifted to super class
Tests are still passing well enough that I'm going to move the
constructors next. The MonadAbstract children are constructor shells
right now.

Once constructors are lifted, the MonadAbstract children will be deleted
and MonadAbstract renamed to Monad. With THAT change we will be back to
where we were many, many years ago, but able to handle a variety of
number types.