monad constructors changed
With the v2.0 approach to tracking coefficients (i.e. not necessarily in
order within Scale) it is necessary to ditch the constructors for Monad
that accept arrays. Now they require Scale's.

That means these constructors are less likely to be used directly in raw
constructions of monads. They are more likely in settings where copying
is occurring.